+++ src/Tests/ContainerGroup.fs
@@ -73,11 +73,6 @@ let tests =
                 Expect.equal group.OsType "Linux" "OS should be Linux"
                 Expect.equal group.IpAddress.Ports.[1].PortProperty 123 "Incorrect udp port"
 
-                Expect.equal
-                    group.NetworkProfile.Id
-                    "[resourceId('Microsoft.Network/networkProfiles', 'test')]"
-                    "Incorrect network profile reference"
-
                 let containerInstance = group.Containers.[0]
                 Expect.equal containerInstance.Image "nginx:1.17.6-alpine" "Incorrect image"
                 Expect.equal containerInstance.Name "nginx" "Incorrect instance name"
@@ -271,10 +266,12 @@ let tests =
             }
 
             test "Container group with private registry" {
+                let managedIdentity = ManagedIdentity.Empty
+
                 let group =
                     containerGroup {
                         add_instances [ nginx ]
-                        add_registry_credentials [ registry "my-registry.azurecr.io" "user" ]
+                        add_registry_credentials [ registry "my-registry.azurecr.io" "user" managedIdentity ]
                     }
                     |> asAzureResource
 
@@ -287,6 +284,51 @@ let tests =
                     credentials.Password
                     "[parameters('my-registry.azurecr.io-password')]"
                     "Container image registry password should be secure parameter"
+            }
+
+            test "Container group with managed identity to private registry" {
+                let userAssignedIdentity =
+                    ResourceId.create (Arm.ManagedIdentity.userAssignedIdentities, ResourceName "user", "resourceGroup")
+                    |> UserAssignedIdentity
+
+                let managedIdentity =
+                    { ManagedIdentity.Empty with
+                        UserAssigned = [ userAssignedIdentity ]
+                    }
+
+                let group =
+                    containerGroup {
+                        add_instances [ nginx ]
+
+                        add_identity (
+                            ResourceId.create (
+                                Arm.ManagedIdentity.userAssignedIdentities,
+                                ResourceName "user",
+                                "resourceGroup"
+                            )
+                            |> UserAssignedIdentity
+                        )
+
+                        add_managed_identity_registry_credentials
+                            [ registry "my-registry.azurecr.io" "user" managedIdentity ]
+                    }
+                    |> asAzureResource
+
+                Expect.hasLength
+                    group.ImageRegistryCredentials
+                    1
+                    "Expected one image managed identity registry credential"
+
+                let credentials = group.ImageRegistryCredentials.[0]
+                Expect.equal credentials.Server "my-registry.azurecr.io" "Incorrect container image registry server"
+                Expect.equal credentials.Username String.Empty "Container image registry user should be null"
+
+                Expect.equal
+                    credentials.Identity
+                    (managedIdentity.Dependencies.Head.ArmExpression.Eval())
+                    "Incorrect container image registry identity"
+
+                Expect.equal credentials.Password null "Container image registry password should be null"
             }
 
             test "Container group with reference to private registry" {
@@ -889,21 +931,7 @@ async {
 
                 Expect.equal (string ipConfigName) "ipconfigProfile" "netprofile ipConfiguration has wrong name"
             }
-            test "Can link a network profile directly to a container group" {
-                let profile = networkProfile { name "netprofile" }
 
-                let template =
-                    containerGroup {
-                        name "appWithHttpFrontend"
-                        network_profile profile
-                    }
-                    |> asAzureResource
-
-                Expect.equal
-                    "[resourceId('Microsoft.Network/networkProfiles', 'netprofile')]"
-                    template.NetworkProfile.Id
-                    "Incorrect profile name"
-            }
             test "Support for additional dependencies" {
                 let storage = storageAccount { name "containerstorage" }